day 01
Change the starter files so that:

Text only comes to the midpoint of the page
The text does not overflow on the bottom on smaller screens
Notes day 1 - Using percentages and avoiding heights
Html is responsive by nature (which doesn't means that it looks great). You can break responsiveness with CSS. If you set a fixed width (like : width: 500px) or height, you can create text-overflow.